Setup automatic version upgrades of pkglist (#8)
Setup the script responsible for upgrading the apt dependencies defined in the provision/pkglist file. This script will run on a weekly frequency, and check to see if any dependencies are available for the base image.

For now the base image is hard-coded to latest, but will later be refactored to use docker run on the base machine (using the base image).